refactor(edge): isolate control api service startup and lifecycle
