+++ crates/edge/src/quic_listener/control_api/mod.rs
@@ -11,7 +11,9 @@ use tokio_rustls::TlsAcceptor;
 use super::*;
 
 mod http;
+mod context;
 mod reload;
+mod service;
 mod state;
 mod watchdog;
